load css files instead of rs files.  default and blue are working.
 have it mostly implemented to switch out themes live,
 but there are some side effects I haven't figured out yet that
 are causing some issues with it, so its disabled for now.
