+++ Makefile.am
@@ -18,8 +18,8 @@ desktop_DATA = jpilot.desktop
 # The skin files for changing jpilot appearance
 colordir = $(pkgdatadir)
 color_DATA = \
-	jpilotrc.blue jpilotrc.default jpilotrc.green jpilotrc.purple \
-	jpilotrc.steel
+	jpilotcss.blue jpilotcss.default jpilotcss.green jpilotcss.purple \
+	jpilotcss.steel
 
 # Instructions for the code to build begins here
