+++ jpilotcss.purple
@@ -5901,7 +5901,7 @@ expander title > arrow:disabled:backdrop {
  * Calendar *
  ***********/
 calendar {
-     color: @theme_unfocused_text_color;
+    color: @theme_unfocused_text_color;
     border: 1px solid @borders;
 }
 calendar.view {
