+++ prefs.c
@@ -20,6 +20,7 @@
 
 /********************************* Includes ***********************************/
 #include "config.h"
+#include "utils.h"
 #include <sys/types.h>
 #include <dirent.h>
 #include <stdio.h>
@@ -49,7 +50,7 @@ static int t_fmt_ampm = TRUE;
 /* These are the default settings */
 /* name, usertype, filetype, ivalue, char *svalue, svalue_size; */
 static prefType glob_prefs[NUM_PREFS] = {
-   {"jpilotrc", CHARTYPE, CHARTYPE, 0, NULL, 0},
+   {"jpilotcss", CHARTYPE, CHARTYPE, 0, NULL, 0},
    {"time", CHARTYPE, INTTYPE, 0, NULL, 0},
    {"sdate", CHARTYPE, INTTYPE, 0, NULL, 0},
    {"ldate", CHARTYPE, INTTYPE, 0, NULL, 0},
@@ -152,12 +153,7 @@ static prefType glob_prefs[NUM_PREFS] = {
    {"external_editor", CHARTYPE, CHARTYPE, 0, NULL, 0},
 };
 
-struct name_list {
-   char *name;
-   struct name_list *next;
-};
 
-static struct name_list *dir_list=NULL;
 
 /****************************** Main Code *************************************/
 void pref_init(void)
@@ -320,86 +316,7 @@ void free_name_list(struct name_list **Plist)
 }
 */
 
-static int get_rcfile_name(int n, char *rc_copy)
-{
-   DIR *dir;
-   struct dirent *dirent;
-   char full_name[FILENAME_MAX];
-   int i;
-   char filename[FILENAME_MAX];
-   int found, count;
-   struct name_list *temp_list, *new_entry;
 
-
-   if (dir_list == NULL) {
-      i = found = count = 0;
-      sprintf(filename, "%s/%s/%s/", BASE_DIR, "share", EPN);
-      jp_logf(JP_LOG_DEBUG, "opening dir %s\n", filename);
-      dir = opendir(filename);
-      if (dir) {
-         for(i=0; (dirent = readdir(dir)); i++) {
-            sprintf(filename, "%s%s", EPN, "rc");
-            if (strncmp(filename, dirent->d_name, strlen(filename))) {
-               continue;
-            } else {
-               jp_logf(JP_LOG_DEBUG, "found %s\n", dirent->d_name);
-               new_entry = malloc(sizeof(struct name_list));
-               if (!new_entry) {
-                  jp_logf(JP_LOG_FATAL, "get_rcfile_name(): %s\n", _("Out of memory"));
-                  return EXIT_FAILURE;
-               }
-               new_entry->name = strdup(dirent->d_name);
-               new_entry->next = dir_list;
-               dir_list = new_entry;
-            }
-         }
-      }
-      if (dir) {
-         closedir(dir);
-      }
-
-      get_home_file_name("", full_name, sizeof(full_name));
-      jp_logf(JP_LOG_DEBUG, "opening dir %s\n", full_name);
-      dir = opendir(full_name);
-      if (dir) {
-         for(; (dirent = readdir(dir)); i++) {
-            sprintf(filename, "%s%s", EPN, "rc");
-            if (strncmp(filename, dirent->d_name, strlen(filename))) {
-               continue;
-            } else {
-               jp_logf(JP_LOG_DEBUG, "found %s\n", dirent->d_name);
-               new_entry = malloc(sizeof(struct name_list));
-               if (!new_entry) {
-                  jp_logf(JP_LOG_FATAL, "get_rcfile_name(): %s 2\n", _("Out of memory"));
-                  return EXIT_FAILURE;
-               }
-               new_entry->name = strdup(dirent->d_name);
-               new_entry->next = dir_list;
-               dir_list = new_entry;
-            }
-         }
-      }
-      if (dir) {
-         closedir(dir);
-      }
-   }
-
-   found = 0;
-   for (i=0, temp_list=dir_list; temp_list; temp_list=temp_list->next, i++) {
-      if (i == n) {
-         g_strlcpy(rc_copy, temp_list->name, MAX_PREF_LEN);
-         found=1;
-         break;
-      }
-   }
-
-   if (found) {
-      return EXIT_SUCCESS;
-   } else {
-      rc_copy[0]='\0';
-      return EXIT_FAILURE;
-   }
-}
 
 /* if n is out of range then this function will fail */
 int get_pref_possibility(int which, int n, char *pref_str)
@@ -714,6 +631,7 @@ int set_pref_possibility(int which, long n, int save)
    r = jp_set_pref(glob_prefs, which, n, str);
    if (save) {
       pref_write_rc_file();
+
    }
 
    if (PREF_CHAR_SET == which)
